requirements.txt
astropy==6.0.1
matplotlib==3.9.2
pyrtcm==1.1.1
rpi_ws281x_mock==0.2.2
